reuse dict builder code
