+++ lindera-dictionary-builder/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod compress;
 pub mod cost_matrix;
+pub mod dict;
 pub mod user_dict;
 
 pub use cost_matrix::{CostMatrixBuilder, CostMatrixBuilderOptions, CostMatrixBuilderOptionsError};
+pub use dict::{DictBuilder, DictBuilderOptions, DictBuilderOptionsError};
 pub use user_dict::{UserDictBuilder, UserDictBuilderOptions, UserDictBuilderOptionsError};
